requirements.txt
annotated-types==0.6.0
anyio==3.7.1
arxiv==2.0.0
astroid==3.0.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
autopep8==2.0.4
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
black==23.11.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
coverage==7.3.2
crashtest==0.4.1
debugpy @ file:///D:/bld/debugpy_1695534453785/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
dulwich==0.21.6
email-validator==2.1.0.post1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastapi==0.104.1
fastjsonschema==2.19.0
feedparser==6.0.10
filelock==3.13.1
flake8==6.1.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
identify==2.5.31
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
iniconfig==2.0.0
installer==0.7.0
ipykernel @ file:///D:/bld/ipykernel_1698244157926/work
ipython @ file:///D:/bld/ipython_1698846796959/work
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///D:/bld/jupyter_core_1698673832306/work
keyring==24.3.0
MarkupSafe==2.1.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mccabe==0.7.0
mongomock==4.1.2
more-itertools==10.1.0
msgpack==1.0.7
mypy==1.7.0
mypy-extensions==1.0.0
nbqa==1.7.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
nodeenv==1.8.0
orjson==3.9.10
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.11.2
pexpect==4.8.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pre-commit==3.5.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1699963054032/work
psutil @ file:///D:/bld/psutil_1695367295750/work
ptyprocess==0.7.0
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycodestyle==2.11.1
pydantic==2.5.1
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.3
pyflakes==3.1.0
Pygments==2.16.1
pylint==3.0.2
pymongo==4.6.0
pyproject_hooks==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
python-multipart==0.0.6
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq @ file:///D:/bld/pyzmq_1698062575483/work
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
sentinels==1.0.0
setuptools==68.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
tokenize-rt==5.2.0
tomli==2.0.1
tomlkit==0.12.3
tornado @ file:///D:/bld/tornado_1695373547508/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698671135544/work
trove-classifiers==2023.11.14
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.24.6
watchfiles==0.21.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1699959196938/work
websockets==12.0
wheel==0.41.3
-e git+https://github.com/schnock-art/arxiv-db.git@20bf98838ebdeeb33c78a9a66f40c42a5522d112#egg=your_project_name
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
